requirements.txt
azure-storage-blob
streamlit
python-dotenv